requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
distlib==0.3.6
Django==4.1
djangorestframework==3.13.1
filelock==3.8.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
virtualenv==20.16.3
